pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.32"
python-dotenv = "^0.19.2"
mariadb = "^1.0.10"

[tool.poetry.scripts]
drive = 'app.orm:main'
fk = 'app.fuckyou:fuck'

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
